+++ defeatbeta_api/data/sql/select_roa_by_industry.sql
@@ -1,13 +1,13 @@
-WITH roe_table AS (
+WITH roa_table AS (
             SELECT
                 symbol,
                 report_date,
                 MAX(CASE WHEN item_name = 'net_income_common_stockholders' THEN item_value END) AS net_income_common_stockholders,
                 MAX(CASE WHEN item_name = 'total_assets' THEN item_value END) AS total_assets
             FROM
-                '{url}'
+                '{stock_statement}'
             WHERE
-                symbol = '{ticker}'
+                symbol in ({symbols})
                 AND item_name IN ('net_income_common_stockholders', 'total_assets')
                 AND report_date != 'TTM'
                 AND period_type = 'quarterly'
@@ -25,7 +25,7 @@ base_data AS (
         QUARTER(report_date::DATE) AS report_quarter,
         YEAR(report_date::DATE) * 4 + QUARTER(report_date::DATE) AS continuous_id
     FROM
-        roe_table
+        roa_table
     WHERE
         net_income_common_stockholders IS NOT NULL AND total_assets IS NOT NULL
 ),
@@ -39,7 +39,7 @@ base_data_rn AS (
         report_year,
         report_quarter,
         continuous_id,
-        ROW_NUMBER() OVER (ORDER BY continuous_id ASC) AS rn_asc
+        ROW_NUMBER() OVER (PARTITION BY symbol ORDER BY continuous_id ASC) AS rn_asc
     FROM
         base_data
 ),
@@ -51,31 +51,17 @@ grouped_data AS (
     FROM
         base_data_rn
 ),
-
 base_data_window AS (
-    SELECT
-        symbol,
-        report_date,
-        net_income_common_stockholders,
-        total_assets,
-        ROW_NUMBER() OVER (ORDER BY report_date ASC) AS rn
-    FROM
-        grouped_data t1
-        JOIN (
-            SELECT
-                group_id
-            FROM
-                grouped_data
-            ORDER BY
-                continuous_id DESC
-                LIMIT 1
-        ) t2
-    ON t1.group_id = t2.group_id
-    ORDER BY
-        continuous_id ASC
+    SELECT *
+    FROM (
+        SELECT
+            *,
+            MAX(group_id) OVER (PARTITION BY symbol) AS max_group_id
+        FROM grouped_data
+    ) t
+    WHERE group_id = max_group_id
 ),
-
-assets_with_lag AS (
+equity_with_lag AS (
     SELECT
         symbol,
         report_date,
@@ -85,30 +71,26 @@ assets_with_lag AS (
     FROM base_data_window
 ),
 
-asserts_avg AS (
+net_incomet_and_avg_asserts AS (
     SELECT
         symbol,
         report_date,
         net_income_common_stockholders,
-        ending_total_assets,
-        beginning_total_assets,
-        (beginning_total_assets + ending_total_assets) / 2.0 AS avg_assets
-    FROM assets_with_lag
+        (beginning_total_assets + ending_total_assets) / 2.0 AS avg_asserts
+    FROM equity_with_lag
     WHERE beginning_total_assets IS NOT NULL
-)
+),
 
-select symbol,
-        report_date,
-        net_income_common_stockholders,
-        beginning_total_assets,
-        ending_total_assets,
-        avg_assets,
-        ROUND(
-            CASE
-                WHEN net_income_common_stockholders < 0 OR avg_assets < 0 THEN
-                    -ABS(net_income_common_stockholders / avg_assets)
-                ELSE
-                    net_income_common_stockholders / avg_assets
-            END
-        , 4) AS roa
-    from asserts_avg order by report_date;
+pivoted AS (
+    SELECT *
+    FROM net_incomet_and_avg_asserts
+    PIVOT (
+        ANY_VALUE(net_income_common_stockholders) AS net_income_common_stockholders,
+        ANY_VALUE(avg_asserts) AS avg_asserts
+        FOR symbol IN ({symbols})
+    )
+)
+SELECT
+    *
+FROM pivoted
+ORDER BY report_date;